% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meansdofcols.R
\name{mean_sd_of_cols}
\alias{mean_sd_of_cols}
\title{Mean and standard deviations of columns.}
\usage{
mean_sd_of_cols(your_data)
}
\arguments{
\item{your_data}{the dataset containing only numeric columns.}
}
\value{
returns a data frame with column means and std. deviations.
}
\description{
Mean and standard deviations of columns.
}
\details{
This function takes the mean and standard deviation of each numeric column in a dataset.
}
\examples{
mean_sd_of_cols(data.frame(a<- c(1,2,3,4), b<- c(0.1,0.2,0.3,0.4)))
}
